% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze.R
\name{analyze_page_server}
\alias{analyze_page_server}
\title{Analysis Page Server}
\usage{
analyze_page_server(id, rv, parent_session)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Reactive values from parent}

\item{parent_session}{Parent session for navigation}
}
\description{
Analysis Page Server
}
\keyword{internal}
