% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze.R
\name{analyze_page_ui}
\alias{analyze_page_ui}
\title{Analysis Page UI}
\usage{
analyze_page_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI elements
}
\description{
Analysis Page UI
}
\keyword{internal}
