% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-main.R
\name{app-main}
\alias{app-main}
\title{canhrQsort Dashboard}
\description{
CANHR-branded Shiny dashboard for Q-sort analysis, following
the canhrActi design system: UAF Blue header, navy sidebar, gold accents.
}
