% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui-theme.R
\name{app-ui-theme}
\alias{app-ui-theme}
\title{canhrQsort Dashboard Theme System}
\description{
Theme for the canhrQsort Shiny dashboard, following the
canhrActi design system: UAF Blue (#236192) + Gold (#FFCD00) branding,
cool slate neutrals, Bootstrap 5 / bslib integration.
}
