% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-main.R
\name{asset_version}
\alias{asset_version}
\title{Asset cache-busting version (file mtime as integer)}
\usage{
asset_version(rel_path)
}
\description{
Asset cache-busting version (file mtime as integer)
}
\keyword{internal}
