% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-bayesian.R
\name{bayesian_page_server}
\alias{bayesian_page_server}
\title{Bayesian Analysis Page Server}
\usage{
bayesian_page_server(id, rv, parent_session)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Reactive values from parent}

\item{parent_session}{Parent session for navigation}
}
\description{
Bayesian Analysis Page Server
}
\keyword{internal}
