% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-bayesian.R
\name{bayesian_page_ui}
\alias{bayesian_page_ui}
\title{Bayesian Analysis Page UI}
\usage{
bayesian_page_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI elements
}
\description{
Bayesian Analysis Page UI
}
\keyword{internal}
