% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Methods for Q-Sort Analysis}
\description{
Bootstrap resampling methods for estimating confidence intervals and
stability of Q-sort analysis results. Based on Zabala & Pascual (2016).
}
\references{
Zabala, A., & Pascual, U. (2016). Bootstrapping Q methodology to improve the
understanding of human perspectives. PloS one, 11(2), e0148087.
}
