% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-bayesian.R
\name{bq_grid_labels}
\alias{bq_grid_labels}
\title{Grid-column index -> conventional score labels (-4..4 style)}
\usage{
bq_grid_labels(C)
}
\description{
Grid-column index -> conventional score labels (-4..4 style)
}
\keyword{internal}
