% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-bayesian.R
\name{bq_unspanned_cluster}
\alias{bq_unspanned_cluster}
\title{Mutual-unspanned cluster signal from a stored person-check table}
\usage{
bq_unspanned_cluster(pc)
}
\description{
Mutual-unspanned cluster signal from a stored person-check table
}
\keyword{internal}
