% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-bayesian.R
\name{bq_write_workbook}
\alias{bq_write_workbook}
\title{Write the complete bayesqm analysis to a multi-sheet Excel workbook}
\usage{
bq_write_workbook(file, bundle, qdata)
}
\description{
Write the complete bayesqm analysis to a multi-sheet Excel workbook
}
\keyword{internal}
