% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-main.R
\name{canhrqsort_server}
\alias{canhrqsort_server}
\title{canhrQsort Dashboard Server}
\usage{
canhrqsort_server(preload_data = NULL)
}
\arguments{
\item{preload_data}{Optional QsortData to preload}
}
\value{
Shiny server function
}
\description{
canhrQsort Dashboard Server
}
\keyword{internal}
