% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compute_bootstrap_ci}
\alias{compute_bootstrap_ci}
\title{Compute Bootstrap Confidence Intervals}
\usage{
compute_bootstrap_ci(samples, original, alpha, type, ...)
}
\description{
Compute Bootstrap Confidence Intervals
}
\keyword{internal}
