% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-core.R
\name{compute_factor_score_se}
\alias{compute_factor_score_se}
\title{Compute Standard Errors of Factor Scores}
\usage{
compute_factor_score_se(sorts, loadings, flags)
}
\arguments{
\item{sorts}{Matrix of Q-sort data}

\item{loadings}{Factor loading matrix}

\item{flags}{Flagging matrix}
}
\value{
Data frame with SE for each factor's scores
}
\description{
Calculate standard errors for factor scores following Brown (1980) and
the PQMethod methodology. The SE depends on:
\enumerate{
\item Number of defining sorts
\item Factor loadings of defining sorts
\item Composite reliability of the factor
}

Formula: SE = sd_zscore * sqrt(1 - reliability)
}
\keyword{internal}
