% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{compute_quality_metrics}
\alias{compute_quality_metrics}
\title{Compute Data Quality Metrics}
\usage{
compute_quality_metrics(data)
}
\arguments{
\item{data}{A QsortData object}
}
\value{
Data frame of quality metrics
}
\description{
Compute data quality metrics for Q-sort data.
}
