% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-plots.R
\name{create_plot_ranking}
\alias{create_plot_ranking}
\title{Create Statement Ranking Plot}
\usage{
create_plot_ranking(results, qdata, factor_num = 1)
}
\arguments{
\item{results}{QsortResults object}

\item{qdata}{QsortData object}

\item{factor_num}{Factor number}
}
\value{
ggplot2 object
}
\description{
Create Statement Ranking Plot
}
