% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-plots.R
\name{create_plot_scree}
\alias{create_plot_scree}
\title{Create Scree Plot}
\usage{
create_plot_scree(results, qdata)
}
\arguments{
\item{results}{QsortResults object}

\item{qdata}{QsortData object}
}
\value{
ggplot2 object
}
\description{
Create Scree Plot
}
