% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demo-data}
\alias{demo-data}
\title{Demo Datasets for Q-Sort Analysis}
\description{
Demonstration datasets for Q-sort methodology analysis. These datasets
are from published research and can be used for testing, learning, and
demonstrating the capabilities of the canhrQsort package.
}
\details{
Both datasets are from:

Akhtar-Danesh, N. (2023). Impact of factor rotation on Q-methodology analysis.
PLOS ONE, 18(9), e0290728. \doi{10.1371/journal.pone.0290728}

The data are available under a Creative Commons Attribution 4.0 International
(CC-BY 4.0) license.
}
\keyword{datasets}
