% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze-factors.R
\name{factors_server_bind}
\alias{factors_server_bind}
\title{Bind the Factors server logic inside the analyze module server}
\usage{
factors_server_bind(input, output, session, rv, ns, config)
}
\description{
Bind the Factors server logic inside the analyze module server
}
\keyword{internal}
