% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-core.R
\name{generate_factor_array}
\alias{generate_factor_array}
\title{Generate Factor Array}
\usage{
generate_factor_array(results, factor_num = 1)
}
\arguments{
\item{results}{A QsortResults object}

\item{factor_num}{Which factor to generate array for}
}
\value{
Data frame with statement positions in the factor array
}
\description{
Generate the factor array (model Q-sort) for a factor. This shows
how the "ideal" participant loading purely on this factor would
sort the statements.
}
\examples{
\dontrun{
array <- generate_factor_array(results, factor_num = 1)
print(array)
}
}
