% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{home_page_server}
\alias{home_page_server}
\title{Overview Page Server}
\usage{
home_page_server(id, rv, parent_session)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Reactive values from parent}

\item{parent_session}{Parent session for navigation}
}
\description{
Overview Page Server
}
\keyword{internal}
