% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{home_page_ui}
\alias{home_page_ui}
\title{Overview Page UI}
\usage{
home_page_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI elements
}
\description{
Overview Page UI
}
\keyword{internal}
