% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{kenq_type1_to_sorts_matrix}
\alias{kenq_type1_to_sorts_matrix}
\title{Convert Ken-Q Type 1 Statement-Number Layout to Canonical Sorts Matrix}
\usage{
kenq_type1_to_sorts_matrix(
  sort_values,
  stmt_data,
  n_participants,
  n_statements,
  participant_names
)
}
\description{
In Type 1, each data row has: column A = sort value, columns B+ = statement
numbers placed at that sort value by each participant. This function inverts
that: for each participant, sorts_matrix\link{p, statement_num} = sort_value.
}
\keyword{internal}
