% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{lipset_qdata}
\alias{lipset_qdata}
\title{The Lipset (1963) benchmark dataset as a QsortData object}
\usage{
lipset_qdata()
}
\description{
The Lipset (1963) benchmark dataset as a QsortData object
}
\keyword{internal}
