% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{load_session_file}
\alias{load_session_file}
\title{Load Session File Helper}
\usage{
load_session_file(file_path, rv, session, parent_session)
}
\description{
Load Session File Helper
}
\keyword{internal}
