% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{ov2_fingerprint}
\alias{ov2_fingerprint}
\title{Build the loaded-data distribution fingerprint in the rank ramp}
\usage{
ov2_fingerprint(distribution)
}
\description{
Build the loaded-data distribution fingerprint in the rank ramp
}
\keyword{internal}
