% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{ov2_outline_pyramid}
\alias{ov2_outline_pyramid}
\title{Build the outline pyramid for the empty drop zone (one gold apex tile)}
\usage{
ov2_outline_pyramid()
}
\description{
Build the outline pyramid for the empty drop zone (one gold apex tile)
}
\keyword{internal}
