% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{ov2_sample_row}
\alias{ov2_sample_row}
\title{One sample dataset row for the empty state}
\usage{
ov2_sample_row(input_id, name, meta, description)
}
\description{
One sample dataset row for the empty state
}
\keyword{internal}
