% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-home.R
\name{ov2_step}
\alias{ov2_step}
\title{One step of the workflow rail}
\usage{
ov2_step(
  title,
  state,
  description = NULL,
  actions = NULL,
  last = FALSE,
  line_done = FALSE
)
}
\description{
One step of the workflow rail
}
\keyword{internal}
