% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_pattern_sheet}
\alias{parse_kenq_pattern_sheet}
\title{Parse Ken-Q Pattern Sheet (20-element multiplier array)}
\usage{
parse_kenq_pattern_sheet(file, sheet_name)
}
\description{
Parse Ken-Q Pattern Sheet (20-element multiplier array)
}
\keyword{internal}
