% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_statements_sheet}
\alias{parse_kenq_statements_sheet}
\title{Parse Ken-Q Statements Sheet}
\usage{
parse_kenq_statements_sheet(file, sheet_name)
}
\description{
Parse Ken-Q Statements Sheet
}
\keyword{internal}
