% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_type1_ver1}
\alias{parse_kenq_type1_ver1}
\title{Parse Ken-Q Type 1 Ver1: Rows 1-22 = header/pattern, Row 23+ = names + data}
\usage{
parse_kenq_type1_ver1(raw, n_statements)
}
\description{
Parse Ken-Q Type 1 Ver1: Rows 1-22 = header/pattern, Row 23+ = names + data
}
\keyword{internal}
