% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_type1_ver2}
\alias{parse_kenq_type1_ver2}
\title{Parse Ken-Q Type 1 Ver2: Row 1 = header/blank + names, Rows 2+ = sort_val + stmt numbers}
\usage{
parse_kenq_type1_ver2(raw, n_statements)
}
\description{
Parse Ken-Q Type 1 Ver2: Row 1 = header/blank + names, Rows 2+ = sort_val + stmt numbers
}
\keyword{internal}
