% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_type2_old}
\alias{parse_kenq_type2_old}
\title{Parse Ken-Q Type 2 Old: has "Sort Pattern" row, project name, then data}
\usage{
parse_kenq_type2_old(raw, sort_pattern_row)
}
\description{
Parse Ken-Q Type 2 Old: has "Sort Pattern" row, project name, then data
}
\keyword{internal}
