% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{parse_kenq_type2_ver2}
\alias{parse_kenq_type2_ver2}
\title{Parse Ken-Q Type 2 Ver2: simple rows of name + direct scores}
\usage{
parse_kenq_type2_ver2(raw)
}
\description{
Parse Ken-Q Type 2 Ver2: simple rows of name + direct scores
}
\keyword{internal}
