% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{participant-attributes}
\alias{participant-attributes}
\title{Participant Attributes}
\description{
Store, match, bin, and retrieve participant attributes
(demographics) for subgroup analysis of Q-sorts
}
