% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot,QsortResults-method}
\alias{plot,QsortResults-method}
\title{Generic Plot Method for QsortResults}
\usage{
\S4method{plot}{QsortResults}(x, y, ...)
}
\arguments{
\item{x}{A QsortResults object}

\item{y}{Not used (included for S4 method signature compatibility)}

\item{...}{Additional arguments passed to plot_qsort}
}
\value{
A ggplot2 plot object
}
\description{
Generic Plot Method for QsortResults
}
