% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot Consensus Statements}
\usage{
plot_consensus(results)
}
\arguments{
\item{results}{A QsortResults object}
}
\value{
A ggplot object
}
\description{
Create a visualization of consensus statements shared across all factors.
}
