% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze-priorities.R
\name{pr_attr_label}
\alias{pr_attr_label}
\title{Priorities body UI (rendered inside the analyze page)}
\usage{
pr_attr_label(x)
}
\description{
Priorities body UI (rendered inside the analyze page)
}
\keyword{internal}
