% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze-priorities.R
\name{priorities_bar_ui}
\alias{priorities_bar_ui}
\title{Priorities control-bar actions (right side of the page bar)}
\usage{
priorities_bar_ui(ns)
}
\description{
Priorities control-bar actions (right side of the page bar)
}
\keyword{internal}
