% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-analyze-priorities.R
\name{priorities_server_bind}
\alias{priorities_server_bind}
\title{Bind the Priorities server logic inside the analyze module server}
\usage{
priorities_server_bind(input, output, session, rv, ns)
}
\description{
Bind the Priorities server logic inside the analyze module server
}
\keyword{internal}
