% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-generic.R
\name{process_long_format}
\alias{process_long_format}
\title{Process Long Format Data}
\usage{
process_long_format(data, participant_col, statement_col, value_col)
}
\description{
Process Long Format Data
}
\keyword{internal}
