% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-core.R
\name{qsort_default_distribution}
\alias{qsort_default_distribution}
\title{Default Q-Sort Distribution}
\usage{
qsort_default_distribution(n_statements)
}
\arguments{
\item{n_statements}{Number of statements}
}
\value{
Vector of counts per position (from most negative to most positive).
For example, c(1, 2, 3, 2, 1) means 1 statement at position -2,
2 at position -1, 3 at position 0, 2 at position 1, 1 at position 2.
}
\description{
Generate a default quasi-normal distribution for a given number of statements.
}
\keyword{internal}
