% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{qsort_results}
\alias{qsort_results}
\title{Create QsortResults Object}
\usage{
qsort_results(
  data,
  correlation,
  extraction,
  rotation,
  flagging,
  factor_scores,
  distinguishing,
  consensus,
  factor_characteristics,
  n_factors,
  method_details,
  call
)
}
\description{
Constructor function for QsortResults S4 class
}
\keyword{internal}
