% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{QsortBootstrap-class}
\alias{QsortBootstrap-class}
\title{QsortBootstrap Class}
\description{
S4 class to store bootstrap results for Q-sort analysis
}
\section{Slots}{

\describe{
\item{\code{original}}{The original QsortResults object}

\item{\code{n_bootstrap}}{Number of bootstrap iterations}

\item{\code{loading_samples}}{Array of bootstrapped factor loadings}

\item{\code{score_samples}}{Array of bootstrapped factor scores}

\item{\code{loading_ci}}{Data frame of confidence intervals for loadings}

\item{\code{score_ci}}{Data frame of confidence intervals for scores}

\item{\code{flag_frequency}}{Matrix of flagging frequencies across bootstraps}

\item{\code{stability_metrics}}{List of stability metrics}

\item{\code{seed}}{Random seed used for reproducibility}
}}

