% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{QsortData-class}
\alias{QsortData-class}
\title{QsortData Class}
\description{
S4 class to store Q-sort data with metadata
}
\section{Slots}{

\describe{
\item{\code{sorts}}{A matrix of Q-sort data (participants x statements)}

\item{\code{statements}}{A character vector of statement texts}

\item{\code{participants}}{A character vector of participant IDs}

\item{\code{distribution}}{A numeric vector defining the forced distribution}

\item{\code{metadata}}{A list containing additional metadata}

\item{\code{source}}{A character string indicating the data source}

\item{\code{created}}{A POSIXct timestamp of when the object was created}
}}

\examples{
\dontrun{
# Create QsortData from matrix
data <- qsort_data(
  sorts = my_matrix,
  statements = paste("Statement", 1:40),
  distribution = c(2, 3, 4, 5, 6, 7, 6, 5, 4, 3, 2)
)
}
}
