% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{QsortResults-class}
\alias{QsortResults-class}
\title{QsortResults Class}
\description{
S4 class to store Q-sort analysis results
}
\section{Slots}{

\describe{
\item{\code{data}}{The original QsortData object}

\item{\code{correlation}}{Correlation matrix between Q-sorts}

\item{\code{extraction}}{List with extraction method details and unrotated loadings}

\item{\code{rotation}}{List with rotation method details and rotated loadings}

\item{\code{flagging}}{Matrix of flagging indicators}

\item{\code{factor_scores}}{Data frame of z-scores for each factor}

\item{\code{distinguishing}}{List of distinguishing statements per factor}

\item{\code{consensus}}{Character vector of consensus statements}

\item{\code{factor_characteristics}}{Data frame of factor characteristics}

\item{\code{n_factors}}{Number of factors extracted}

\item{\code{method_details}}{List of method parameters used}

\item{\code{call}}{The original function call}
}}

