% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{range_lower_bound}
\alias{range_lower_bound}
\title{Lower Bound of a Range Level}
\usage{
range_lower_bound(x)
}
\description{
Parse the leading number of a range string for ordering.
}
\keyword{internal}
