% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{read_kenq}
\alias{read_kenq}
\title{Read Q-Sort Data from Ken-Q Analysis}
\usage{
read_kenq(file, format = c("auto", "json", "csv"))
}
\arguments{
\item{file}{Path to Ken-Q export file (JSON or CSV)}

\item{format}{Character; export format - "auto", "json", or "csv"}
}
\value{
A QsortData object
}
\description{
Import Q-sort data from Ken-Q Analysis export files. Ken-Q exports
data in JSON or CSV format.
}
\examples{
\dontrun{
qdata <- read_kenq("kenq_export.json")
qdata <- read_kenq("kenq_export.csv")
}
}
\references{
Banasick, S. Ken-Q Analysis.
\url{https://shawnbanasick.github.io/ken-q-analysis/}
}
