% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{read_pqmethod}
\alias{read_pqmethod}
\title{Read Q-Sort Data from PQMethod}
\usage{
read_pqmethod(file, statements_file = NULL)
}
\arguments{
\item{file}{Path to the PQMethod .DAT file}

\item{statements_file}{Optional path to separate statements file (.STA)}
}
\value{
A QsortData object
}
\description{
Import Q-sort data from PQMethod .DAT files. PQMethod is a widely used
DOS-based program for Q methodology analysis.
}
\examples{
\dontrun{
qdata <- read_pqmethod("lipset.dat")
qdata <- read_pqmethod("study.dat", statements_file = "study.sta")
}
}
\references{
Schmolck, P. (2014). PQMethod (Version 2.35).
\url{http://schmolck.org/qmethod/}
}
