% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{read_pqmethod_statements}
\alias{read_pqmethod_statements}
\title{Read PQMethod Statements File}
\usage{
read_pqmethod_statements(file)
}
\description{
Read PQMethod Statements File
}
\keyword{internal}
