% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{run_bootstrap_iteration}
\alias{run_bootstrap_iteration}
\title{Run Single Bootstrap Iteration}
\usage{
run_bootstrap_iteration(
  boot_sorts,
  nfactors,
  extraction,
  rotation,
  target_loadings,
  original_rownames
)
}
\description{
Run Single Bootstrap Iteration
}
\keyword{internal}
