% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui-theme.R
\name{setup_thematic}
\alias{setup_thematic}
\title{Setup Thematic}
\usage{
setup_thematic(mode = "light")
}
\arguments{
\item{mode}{Theme mode (kept for backward compatibility)}
}
\description{
Setup Thematic
}
\keyword{internal}
