% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,QsortData-method}
\alias{show,QsortData-method}
\title{Show Method for QsortData}
\usage{
\S4method{show}{QsortData}(object)
}
\arguments{
\item{object}{A QsortData object}
}
\description{
Show Method for QsortData
}
