% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,QsortResults-method}
\alias{show,QsortResults-method}
\title{Show Method for QsortResults}
\usage{
\S4method{show}{QsortResults}(object)
}
\arguments{
\item{object}{A QsortResults object}
}
\description{
Show Method for QsortResults
}
