% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary,QsortBootstrap-method}
\alias{summary,QsortBootstrap-method}
\title{Summary Method for QsortBootstrap}
\usage{
\S4method{summary}{QsortBootstrap}(object)
}
\arguments{
\item{object}{A QsortBootstrap object}
}
\value{
Invisibly returns the bootstrap object
}
\description{
Summary Method for QsortBootstrap
}
