% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary,QsortResults-method}
\alias{summary,QsortResults-method}
\title{Summary Method for QsortResults}
\usage{
\S4method{summary}{QsortResults}(object)
}
\arguments{
\item{object}{A QsortResults object}
}
\description{
Summary Method for QsortResults
}
