% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-specialized.R
\name{try_parse_firebase_json}
\alias{try_parse_firebase_json}
\title{Try Parsing as Easy-HTMLQ Firebase JSON}
\usage{
try_parse_firebase_json(data)
}
\description{
Try Parsing as Easy-HTMLQ Firebase JSON
}
\keyword{internal}
