% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-visualize.R
\name{visualize_page_server}
\alias{visualize_page_server}
\title{Visualization Page Server}
\usage{
visualize_page_server(id, rv, parent_session)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Reactive values from parent}

\item{parent_session}{Parent session for navigation}
}
\description{
Visualization Page Server
}
\keyword{internal}
