% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-page-visualize.R
\name{visualize_page_ui}
\alias{visualize_page_ui}
\title{Visualization Page UI}
\usage{
visualize_page_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI elements
}
\description{
Visualization Page UI
}
\keyword{internal}
